feat: Separate page into components
