+++ components/Statistics.vue
@@ -1,0 +1,102 @@
+<template>
+  <h2
+    class="text-center font-header text-4xl font-semibold uppercase text-primary sm:text-5xl lg:text-6xl"
+  >
+    My record
+  </h2>
+  <div class="container">
+    <div
+      class="mx-auto w-5/6 bg-white py-16 shadow md:w-11/12 lg:py-20 xl:py-24 2xl:w-full"
+    >
+      <div class="grid grid-cols-2 gap-5 md:gap-8 xl:grid-cols-4 xl:gap-5">
+        <div
+          class="flex flex-col items-center justify-center text-center md:flex-row md:text-left"
+        >
+          <div>
+            <img
+              src="/assets/img/icon-project.svg"
+              class="mx-auto h-12 w-auto md:h-20"
+              alt="icon project"
+            />
+          </div>
+          <div class="pt-5 md:pl-5 md:pt-0">
+            <h1 class="font-body text-2xl font-bold text-primary md:text-4xl">
+              12
+            </h1>
+            <h4
+              class="text-grey-dark font-header text-base font-medium leading-loose md:text-xl"
+            >
+              Finished Projects
+            </h4>
+          </div>
+        </div>
+
+        <div
+          class="flex flex-col items-center justify-center text-center md:flex-row md:text-left"
+        >
+          <div>
+            <img
+              src="/assets/img/icon-award.svg"
+              class="mx-auto h-12 w-auto md:h-20"
+              alt="icon award"
+            />
+          </div>
+          <div class="pt-5 md:pl-5 md:pt-0">
+            <h1 class="font-body text-2xl font-bold text-primary md:text-4xl">
+              3
+            </h1>
+            <h4
+              class="text-grey-dark font-header text-base font-medium leading-loose md:text-xl"
+            >
+              Awards Won
+            </h4>
+          </div>
+        </div>
+
+        <div
+          class="mt-6 flex flex-col items-center justify-center text-center md:mt-10 md:flex-row md:text-left lg:mt-0"
+        >
+          <div>
+            <img
+              src="/assets/img/icon-happy.svg"
+              class="mx-auto h-12 w-auto md:h-20"
+              alt="icon happy clients"
+            />
+          </div>
+          <div class="pt-5 md:pl-5 md:pt-0">
+            <h1 class="font-body text-2xl font-bold text-primary md:text-4xl">
+              8
+            </h1>
+            <h4
+              class="text-grey-dark font-header text-base font-medium leading-loose md:text-xl"
+            >
+              Happy Clients
+            </h4>
+          </div>
+        </div>
+
+        <div
+          class="mt-6 flex flex-col items-center justify-center text-center md:mt-10 md:flex-row md:text-left lg:mt-0"
+        >
+          <div>
+            <img
+              src="/assets/img/icon-puzzle.svg"
+              class="mx-auto h-12 w-auto md:h-20"
+              alt="icon puzzle"
+            />
+          </div>
+          <div class="pt-5 md:pl-5 md:pt-0">
+            <h1 class="font-body text-2xl font-bold text-primary md:text-4xl">
+              99
+            </h1>
+            <h4
+              class="text-grey-dark font-header text-base font-medium leading-loose md:text-xl"
+            >
+              Bugs Fixed
+            </h4>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
